+++ .eslintrc.js
@@ -7,17 +7,18 @@ module.exports = {
   extends: ['airbnb-base'],
   globals: {
     Atomics: 'readonly',
-    SharedArrayBuffer: "readonly"
+    SharedArrayBuffer: 'readonly'
   },
   parserOptions: {
     ecmaVersion: 2018,
     sourceType: 'module'
   },
   rules: {
-    "no-await-in-loop": 'off',
-    "no-restricted-syntax": 'off',
-    "no-console": "off",
+    'no-await-in-loop': 'off',
+    'no-restricted-syntax': 'off',
+    'no-console': 'off',
+    'no-unused-vars': ['warning', {vars: 'local'}],
     quotes: ['error', 'single', { avoidEscape: false, allowTemplateLiterals: true }],
-    "comma-dangle": ['error', 'always-multiline']
+    'comma-dangle': ['error', 'always-multiline']
   }
 };
