+++ index.js
@@ -6,7 +6,7 @@
 
 const debug = require('debug')('a:server');
 const http = require('http');
-const app = require('./app');
+const app = require('./src/app');
 
 require('dotenv').config({ path: '.env' });
