+++ src/api/auth/login.js
@@ -2,7 +2,7 @@ const express = require('express');
 const { body, validationResult } = require('express-validator');
 const bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
-const { buildRes } = require('../../utils/response');
+const { buildRes } = require('../../helper/utils/response');
 const { Users } = require('../../models');
 
 const jwtPrivateKey = process.env.PRIVATE_KEY_JWT || '!bE8JX7!owd!W67&XEU9kw2W';
