+++ src/models/index.js
@@ -15,15 +15,15 @@ const FilmCategoryModel = require('./many_to_many/films_categories');
 const FilmListModel = require('./many_to_many/films_lists');
 
 // Initial data for dev purpose
-const migrationFilmList = require('../db_data/films_data');
-const migrationEpList = require('../db_data/episode_data');
-const migrationProcess = require('../db_data/progress_data');
-const migrationUser = require('../db_data/user_data');
-const migrationActor = require('../db_data/actor_data');
-const migrationCategory = require('../db_data/category_data');
-const migrationList = require('../db_data/list_data');
-const migrationFilmActor = require('../db_data/film_actor_data');
-const migrationFilmCategory = require('../db_data/film_category_data');
+const migrationFilmList = require('../mock/films_data');
+const migrationEpList = require('../mock/episode_data');
+const migrationProcess = require('../mock/progress_data');
+const migrationUser = require('../mock/user_data');
+const migrationActor = require('../mock/actor_data');
+const migrationCategory = require('../mock/category_data');
+const migrationList = require('../mock/list_data');
+const migrationFilmActor = require('../mock/film_actor_data');
+const migrationFilmCategory = require('../mock/film_category_data');
 
 
 const DATABASE_NAME = process.env.DATABASE_NAME || 'math_app';
